move logError to diagnostic
