+++ packages/http-client-java/emitter/src/emitter.ts
@@ -1,123 +1,19 @@
-import {
-  createTypeSpecLibrary,
-  EmitContext,
-  getNormalizedAbsolutePath,
-  JSONSchemaType,
-  resolvePath,
-} from "@typespec/compiler";
+import { $lib } from "@azure-tools/typespec-azure-core";
+import { EmitContext, getNormalizedAbsolutePath, NoTarget, resolvePath } from "@typespec/compiler";
 import { promises } from "fs";
 import { dump } from "js-yaml";
 import { dirname } from "path";
 import { fileURLToPath } from "url";
 import { CodeModelBuilder } from "./code-model-builder.js";
 import { CodeModel } from "./common/code-model.js";
-import { logError, spawnAsync, SpawnError, trace } from "./utils.js";
-import {
-  CODE_JAVA_SDK_DEPENDENCY,
-  JDK_NOT_FOUND_MESSAGE,
-  validateDependencies,
-} from "./validate.js";
-
-export interface EmitterOptions {
-  namespace?: string;
-  "package-dir"?: string;
-
-  flavor?: string;
-
-  "service-name"?: string;
-  "service-versions"?: string[];
-
-  "skip-special-headers"?: string[];
-
-  "generate-samples"?: boolean;
-  "generate-tests"?: boolean;
-
-  "enable-sync-stack"?: boolean;
-  "stream-style-serialization"?: boolean;
-  "use-object-for-unknown"?: boolean;
-
-  "partial-update"?: boolean;
-  "models-subpackage"?: string;
-  "custom-types"?: string;
-  "custom-types-subpackage"?: string;
-  "customization-class"?: string;
-  polling?: any;
-
-  "group-etag-headers"?: boolean;
-
-  "enable-subclient"?: boolean;
-
-  "advanced-versioning"?: boolean;
-  "api-version"?: string;
-  "service-version-exclude-preview"?: boolean;
-
-  "dev-options"?: DevOptions;
-}
-
-export interface DevOptions {
-  "generate-code-model"?: boolean;
-  debug?: boolean;
-  loglevel?: "off" | "debug" | "info" | "warn" | "error";
-  "java-temp-dir"?: string; // working directory for java codegen, e.g. transformed code-model file
-}
+import { EmitterOptions, reportDiagnostic } from "./lib.js";
+import { spawnAsync, SpawnError, trace } from "./utils.js";
+import { validateDependencies } from "./validate.js";
 
 type CodeModelEmitterOptions = EmitterOptions & {
   "output-dir": string;
   arm?: boolean;
 };
-
-const EmitterOptionsSchema: JSONSchemaType<EmitterOptions> = {
-  type: "object",
-  additionalProperties: true,
-  properties: {
-    namespace: { type: "string", nullable: true },
-    "package-dir": { type: "string", nullable: true },
-
-    flavor: { type: "string", nullable: true },
-
-    // service
-    "service-name": { type: "string", nullable: true },
-    "service-versions": { type: "array", items: { type: "string" }, nullable: true },
-
-    // header
-    "skip-special-headers": { type: "array", items: { type: "string" }, nullable: true },
-
-    // sample and test
-    "generate-samples": { type: "boolean", nullable: true, default: true },
-    "generate-tests": { type: "boolean", nullable: true, default: true },
-
-    "enable-sync-stack": { type: "boolean", nullable: true, default: true },
-    "stream-style-serialization": { type: "boolean", nullable: true, default: true },
-    "use-object-for-unknown": { type: "boolean", nullable: true, default: false },
-
-    // customization
-    "partial-update": { type: "boolean", nullable: true, default: false },
-    "models-subpackage": { type: "string", nullable: true },
-    "custom-types": { type: "string", nullable: true },
-    "custom-types-subpackage": { type: "string", nullable: true },
-    "customization-class": { type: "string", nullable: true },
-    polling: { type: "object", additionalProperties: true, nullable: true },
-
-    "group-etag-headers": { type: "boolean", nullable: true },
-
-    "enable-subclient": { type: "boolean", nullable: true, default: false },
-
-    "advanced-versioning": { type: "boolean", nullable: true, default: false },
-    "api-version": { type: "string", nullable: true },
-    "service-version-exclude-preview": { type: "boolean", nullable: true, default: false },
-
-    "dev-options": { type: "object", additionalProperties: true, nullable: true },
-  },
-  required: [],
-};
-
-export const $lib = createTypeSpecLibrary({
-  name: "@typespec/http-client-java",
-  diagnostics: {},
-  emitter: {
-    options: EmitterOptionsSchema,
-  },
-});
 
 export async function $onEmit(context: EmitContext<EmitterOptions>) {
   const program = context.program;
@@ -138,7 +34,11 @@ export async function $onEmit(context: EmitContext<EmitterOptions>) {
       const builder = new CodeModelBuilder(program, context);
       codeModel = await builder.build();
     } catch (error: any) {
-      logError(program, error.message);
+      reportDiagnostic(program, {
+        code: "unknown-error",
+        format: { errorMessage: error.message },
+        target: NoTarget,
+      });
       trace(program, error.stack);
     }
 
@@ -158,7 +58,11 @@ export async function $onEmit(context: EmitContext<EmitterOptions>) {
 
       await promises.mkdir(outputPath, { recursive: true }).catch((err) => {
         if (err.code !== "EISDIR" && err.code !== "EEXIST") {
-          logError(program, `Failed to create output directory: ${outputPath}`);
+          reportDiagnostic(program, {
+            code: "unknown-error",
+            format: { errorMessage: `Failed to create output directory: ${outputPath}.` },
+            target: NoTarget,
+          });
           return;
         }
       });
@@ -198,12 +102,19 @@ export async function $onEmit(context: EmitContext<EmitterOptions>) {
         trace(program, `Code generation log: ${result.stdout}`);
       } catch (error: any) {
         if (error && "code" in error && error["code"] === "ENOENT") {
-          logError(program, JDK_NOT_FOUND_MESSAGE, CODE_JAVA_SDK_DEPENDENCY);
+          reportDiagnostic(program, {
+            code: "invalid-java-sdk-dependency",
+            target: NoTarget,
+          });
         } else {
-          logError(
-            program,
-            'The emitter was unable to generate client code from this TypeSpec, please run this command again with "--trace http-client-java" to get diagnostic information, and open an issue on https://github.com/microsoft/typespec',
-          );
+          reportDiagnostic(program, {
+            code: "unknown-error",
+            format: {
+              errorMessage:
+                'The emitter was unable to generate client code from this TypeSpec, please run this command again with "--trace http-client-java" to get diagnostic information, and open an issue on https://github.com/microsoft/typespec',
+            },
+            target: NoTarget,
+          });
           if (error instanceof SpawnError) {
             trace(program, `Code generation error: ${error.stdout}`);
           }
