+++ packages/http-client-java/emitter/src/utils.ts
@@ -1,15 +1,6 @@
 import { NoTarget, Program, Type } from "@typespec/compiler";
 import { spawn, SpawnOptions } from "child_process";
 
-export function logError(program: Program, msg: string, code: string = "http-client-java") {
-  program.reportDiagnostic({
-    code: code,
-    severity: "error",
-    message: msg,
-    target: NoTarget,
-  });
-}
-
 export function logWarning(program: Program, msg: string, code: string = "http-client-java") {
   program.reportDiagnostic({
     code: code,
